+++ tests/misc_tests/log_manager_test.py
@@ -6,8 +6,8 @@ from os.path import isfile, join
 from time import sleep
 
 from uuid import uuid4
-from log_manager.log_writer import LogWriter
-from log_manager.log_reader import LogReader
+from file_manager.log_writer import LogWriter
+from file_manager.log_reader import LogReader
 import config
 
 # Prefix for files generated by this function
