+++ tests/misc_tests/team_reader_test.py
@@ -1,6 +1,6 @@
 """Testing file for TeamReader."""
 
-from log_manager.team_reader import TeamReader
+from file_manager.team_reader import TeamReader
 
 def test_init():
     """Test initialization method for TeamReader."""
